pyproject.toml
[tool.poetry]
name = "fast_oauth"
version = "0.1.0"
description = ""
authors = ["mabittar <ma_bittar@yaho.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.93.0"
black = "^23.1.0"
isort = "^5.12.0"
autoflake = "^2.0.1"
flake8 = "^6.0.0"
pre-commit = "^3.1.1"
python-multipart = "^0.0.6"
uvicorn = "^0.20.0"
pre-commit-hooks = "^4.4.0"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"
requests-mock = "^1.10.0"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json", ".rst"]
skip_glob = ["docs/*", "*/__init__.py"]
